checkpoint: moves "*/FetchingError.ts" to "*/Fetching/Error.ts"
